Added in Fade in and out. Reworked tutorial puzzle
Troubleshooted a ton of things. Many of these scrips will likely be deleted or never used. Pulled elements out of parts but ended up simplifying then the tutorial that I used.

Still needing the ability to queue up dialogue again based on events in game. Still working on how to do that.